+++ Kata.Checkout/ProductCatalogue/IItem.cs
@@ -3,5 +3,5 @@
 public interface IItem
 {
     public string Name { get; }
-    public float CalculatePrice(int quantity) => 0;
+    public int CalculatePrice(int quantity) => 0;
 }
