+++ Kata.Checkout/ProductCatalogue/MultiBuyItem.cs
@@ -1,8 +1,8 @@
 ﻿namespace Kata.Checkout.Items;
 
-public class MultiBuyItem(float unitPrice, int discountQuantity, float discountPrice) : Item(unitPrice)
+public class MultiBuyItem(int unitPrice, int discountQuantity, int discountPrice) : Item(unitPrice)
 {
-    public override float CalculatePrice(int quantity)
+    public override int CalculatePrice(int quantity)
     {
         if (quantity > discountQuantity)
             return base.CalculatePrice(quantity);
